fix(funnels): widen funnel UDF steps bitfield to UInt64
Track reached funnel steps in a u64 so up to 64 steps can register, and fix the unordered-funnel mask overflowing at full completion. The dev ClickHouse schema now declares UInt64; binaries must be rebuilt with funnel-udf/build.sh before merge. MAX_FUNNEL_STEPS stays 32 until a UDF version built from this source is deployed and UDF_VERSION points at it.

Generated-By: PostHog Desktop
Task-Id: 5b91d4f1-f886-4bf7-bbb5-9c21db12a638

+++ funnel-udf/src/steps.rs
@@ -13,7 +13,7 @@ pub struct EnteredTimestamp {
     pub excluded: bool,
     pub timings: Vec<f64>,
     pub uuids: Vec<Uuid>,
-    pub steps: u32,
+    pub steps: u64,
 }
 
 #[derive(Debug, Clone, Deserialize)]
@@ -42,7 +42,7 @@ pub struct Result(
     pub PropVal,
     pub Vec<f64>,
     pub Vec<Vec<Uuid>>,
-    pub u32,
+    pub u64,
 );
 
 struct Vars {
